Bugfix and sanity checks for counts:
Bugfix:
An array property of wanted name still is available in the count of the parent node.
The child node of that node are available more often!

Sanity checks:
If the next mask key starts with the previous mask key, then the next count must be <= the previous count
If the next mask key does not start with the previous mask key, then the count must be smaller than the first/root count!