Properties with tag 'notDisplayed' can not be masked and restored!
Getting right when to call recursively though property has tag 'notDisplayed' is not trivial!
We need all property names.
We do not need the associated mask keys in loop of method testMaskOne(), but it would have been nice
to have them all to populate maskKey2propName2deleteCount. Therefore add new inner maps as necessary
on the fly in closure addCount2!